refactor: 검색 폴더 domain으로 이동 (#65)
