+++ src/main/java/com/api/ttoklip/domain/search/controller/SearchController.java
@@ -1,10 +1,10 @@
-package com.api.ttoklip.domain.common.search.controller;
+package com.api.ttoklip.domain.search.controller;
 
-import com.api.ttoklip.domain.common.search.constant.SearchResponseConstant;
-import com.api.ttoklip.domain.common.search.response.CommunityPaging;
-import com.api.ttoklip.domain.common.search.response.HoneyTipPaging;
-import com.api.ttoklip.domain.common.search.response.NewsletterPaging;
-import com.api.ttoklip.domain.common.search.service.SearchService;
+import com.api.ttoklip.domain.search.constant.SearchResponseConstant;
+import com.api.ttoklip.domain.search.response.CommunityPaging;
+import com.api.ttoklip.domain.search.response.HoneyTipPaging;
+import com.api.ttoklip.domain.search.response.NewsletterPaging;
+import com.api.ttoklip.domain.search.service.SearchService;
 import com.api.ttoklip.global.success.SuccessResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
